feat(service): support git url builds
